app/requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
sqlalchemy-utils
plotly.express
numpy
pydeck